First batch of changes from PR review comments
