+++ plugins/chain_plugin/chain_plugin.cpp
@@ -1832,56 +1832,6 @@ auto make_resolver(const controller& control, abi_serializer::yield_function_t y
    };
 }
 
-class abi_serializer_cache_builder {
-public:
-   abi_serializer_cache_builder(const controller& db, const fc::microseconds& max_time) :
-      resolver(make_resolver(db, abi_serializer::create_yield_function(max_time)))
-   {
-   }
-
-   abi_serializer_cache_builder(const abi_serializer_cache_builder&) = delete;
-
-   abi_serializer_cache_builder&& add_serializers(const chain::signed_block_ptr& block) && {
-      for( const auto& receipt: block->transactions ) {
-         if( std::holds_alternative<chain::packed_transaction>( receipt.trx ) ) {
-            const auto& pt = std::get<chain::packed_transaction>( receipt.trx );
-            const auto& t = pt.get_transaction();
-            for( const auto& a: t.actions )
-               add_to_cache( a );
-            for( const auto& a: t.context_free_actions )
-               add_to_cache( a );
-         }
-      }
-      return std::move(*this);
-   }
-
-   abi_serializer_cache_builder&& add_serializers(const transaction_trace_ptr& trace_ptr) && {
-      for( const auto& trace: trace_ptr->action_traces ) {
-         add_to_cache(trace.act);
-      }
-      return std::move(*this);
-   }
-
-   chain_apis::abi_serializer_cache&& get() && {
-      return std::move(abi_cache);
-   }
-
-private:
-   void add_to_cache(const chain::action& a) {
-      auto it = abi_cache.find( a.account );
-      if( it == abi_cache.end() ) {
-         try {
-            abi_cache.emplace_hint( it, a.account, resolver( a.account ) );
-         } catch( ... ) {
-            // keep behavior of not throwing on invalid abi, will result in hex data
-         }
-      }
-   }
-
-   std::function<std::optional<abi_serializer> (const account_name &name)> resolver;
-   chain_apis::abi_serializer_cache abi_cache;
-};
-
 read_only::get_scheduled_transactions_result
 read_only::get_scheduled_transactions( const read_only::get_scheduled_transactions_params& p, const fc::time_point& deadline ) const {
 
@@ -1988,8 +1938,9 @@ chain::signed_block_ptr read_only::get_raw_block(const read_only::get_raw_block_
 
 std::function<chain::t_or_exception<fc::variant>()> read_only::get_block(const get_raw_block_params& params, const fc::time_point& deadline) const {
    chain::signed_block_ptr block = get_raw_block(params, deadline);
-   
-   auto abi_cache = abi_serializer_cache_builder(db, deadline - fc::time_point::now()).add_serializers(block).get();
+
+   auto yield = abi_serializer::create_yield_function(deadline - fc::time_point::now());
+   auto abi_cache = abi_serializer_cache_builder(make_resolver(db, std::move(yield))).add_serializers(block).get();
    FC_CHECK_DEADLINE(deadline);
 
    using return_type = t_or_exception<fc::variant>;
@@ -2046,7 +1997,8 @@ read_only::get_block_header_result read_only::get_block_header(const read_only::
 
 abi_resolver
 read_only::get_block_serializers( const chain::signed_block_ptr& block, const fc::microseconds& max_time ) const {
-   return abi_resolver(abi_serializer_cache_builder(db, max_time).add_serializers(block).get());
+   auto yield = abi_serializer::create_yield_function(max_time);
+   return abi_resolver(abi_serializer_cache_builder(make_resolver(db, std::move(yield))).add_serializers(block).get());
 }
 
 fc::variant read_only::convert_block( const chain::signed_block_ptr& block,
@@ -2261,7 +2213,8 @@ void read_write::send_transaction(const read_write::send_transaction_params& par
             next(std::get<fc::exception_ptr>(result));
          } else {
             auto trx_trace_ptr = std::get<transaction_trace_ptr>(result);
-            auto abi_cache     =  abi_serializer_cache_builder(db, fc::microseconds::maximum()).add_serializers(trx_trace_ptr).get();
+            auto yield         = abi_serializer::create_yield_function(fc::microseconds::maximum());
+            auto abi_cache     =  abi_serializer_cache_builder(make_resolver(db, std::move(yield))).add_serializers(trx_trace_ptr).get();
             using return_type  = t_or_exception<read_write::send_transaction_results>;
             next([this, trx_trace_ptr, resolver = abi_resolver(std::move(abi_cache))]() mutable {
                try {
@@ -2320,7 +2273,8 @@ void read_write::send_transaction2(const read_write::send_transaction2_params& p
                            }
                         } );
                   } else {
-                     auto abi_cache    =  abi_serializer_cache_builder(db, fc::microseconds::maximum()).add_serializers(trx_trace_ptr).get();
+                     auto yield        = abi_serializer::create_yield_function(fc::microseconds::maximum());
+                     auto abi_cache    =  abi_serializer_cache_builder(make_resolver(db, std::move(yield))).add_serializers(trx_trace_ptr).get();
                      using return_type = t_or_exception<read_write::send_transaction_results>;
                      next([this, trx_trace_ptr, resolver = abi_resolver(std::move(abi_cache))]() mutable {
                         try {
@@ -2635,7 +2589,8 @@ void read_only::send_transient_transaction(const Params& params, next_function<R
                    next(std::get<fc::exception_ptr>(result));
               } else {
                  auto trx_trace_ptr = std::get<transaction_trace_ptr>(result);
-                 auto abi_cache     =  abi_serializer_cache_builder(db, fc::microseconds::maximum()).add_serializers(trx_trace_ptr).get();
+                 auto yield         = abi_serializer::create_yield_function(fc::microseconds::maximum());
+                 auto abi_cache     =  abi_serializer_cache_builder(make_resolver(db, std::move(yield))).add_serializers(trx_trace_ptr).get();
                  using return_type = t_or_exception<Results>;
                  next([this, trx_trace_ptr, resolver = abi_resolver(std::move(abi_cache))]() mutable {
                     try {
